Move SFTExt to seed_pallet_common
