+++ pallet/marketplace/Cargo.toml
@@ -22,7 +22,6 @@ scale-info = { version = "2.3.0", default-features = false, features = ["derive"
 seed-primitives = { path = "../../primitives", default-features = false }
 seed-pallet-common = { path = "../common", default-features = false }
 pallet-nft = { path = "../nft", default-features = false }
-pallet-sft = { path = "../sft", default-features = false }
 log = { default-features = false, version = "0.4.14"}
 
 [dev-dependencies]
@@ -33,6 +32,7 @@ pallet-balances = { git = "https://github.com/paritytech/substrate", branch = "p
 pallet-assets = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.30" }
 pallet-evm = { git = "https://github.com/futureversecom/frontier", branch = "polkadot-v0.9.30-TRN" }
 pallet-timestamp = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.30" }
+pallet-sft = { path = "../sft" }
 
 [features]
 default = ["std"]
@@ -40,7 +40,6 @@ std = [
 	"seed-primitives/std",
 	"seed-pallet-common/std",
 	"pallet-nft/std",
-	"pallet-sft/std",
 	"codec/std",
 	"serde",
 	"sp-core/std",
